requirements.txt
asgiref==3.3.1
Django==3.1.4
django-environ==0.4.5
django-extensions==3.1.0
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==5.3.0
gunicorn==20.0.4
mysql-connector-python==8.0.22
protobuf==3.14.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
tblib==1.7.0
text-unidecode==1.3
Werkzeug==1.0.1
